e11y first pass: move body basics to separate component
